+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHJN2DJ0EN4RCWQAWW6QZ7GH"
+        "01KHJSX6FMBBB0S07QR0R5T0SJ"
     )
     r"""The unique identifier of the data source"""