+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHJN2DHV7QAX3EJJ92W14SYC"
+        "01KHJSX6FH4JANRPERXA29SN6D"
     )
     r"""The unique identifier of the data source"""