+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHJN2DHYH0YSE0NW9AVGW7SA"
+        "01KHJSX6FKTVTVKT868S5ERV7Y"
     )
     r"""The unique identifier of the data source"""