+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-16T07:18:52.260Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-16T08:43:25.167Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")