+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHJN2DJ4NSFF3DRBVNVW4AJK"
+        "01KHJSX6FP35M69D5AH2ZP2360"
     )
     r"""The unique identifier of the data source"""